Update pkg/podkiller tests
This updates the tests for pkg/podkiller to be more reliable by only checking that the expected number of pods were killed once rather than in between each kill. This makes the test less susceptible to race conditions.

This also adds a call to seed the random number generator to init() for podkiller, which was missing before.